model_api/requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
psycopg2-binary
# To add security
typing_extensions
python-multipart
python-jose[cryptography]
passlib[bcrypt]
scikit-learn
xgboost